+++ SourceCode/include/tracer/Triangle.h
@@ -1,16 +1,16 @@
 #pragma once
 #include <array>
 
-#include "Ray.h"
+#include "Vector.h"
 
-#define DIMENSION 3
+#define TRIANGLE_NUM_VERTICES 3
 class Triangle {
  private:
-  std::array<Vector, DIMENSION> vertices;
+  std::array<Vector, TRIANGLE_NUM_VERTICES> vertices;
 
  public:
   Triangle();
   explicit Triangle(std::array<Vector, 3> &vertices);
-  Vector normalVector() const;
+  Vector getTriangleNormal() const;
   float area() const;
 };
