+++ SourceCode/src/CMakeLists.txt
@@ -1,0 +1,11 @@
+# Get all header files
+file(GLOB HEADER_LIST CONFIGURE_DEPENDS "${RayTracer_SOURCE_DIR}/include/tracer/*.h")
+# Make a library using the sources and the headers
+add_library(tracer Vector.cpp Ray.cpp Triangle.cpp RayTracer.cpp ${HEADER_LIST})
+# This directory is needed in order to access the headers
+target_include_directories(tracer PUBLIC ../include)
+# Make grouping pretty in IDEs
+source_group(
+    TREE "${PROJECT_SOURCE_DIR}/include"
+    PREFIX "Header Files"
+    FILES ${HEADER_LIST})
