+++ SourceCode/src/Vector.cpp
@@ -1,9 +1,8 @@
-#include "Ray.h"
+#include <tracer/Vector.h>
 
 #include <cmath>
-#include <ostream>
 
-Vector::Vector() : x(0), y(0), z(0){};
+Vector::Vector() : x(0.0f), y(0.0f), z(0.0f){};
 
 Vector::Vector(float x, float y, float z) : x(x), y(y), z(z){};
 
@@ -44,17 +43,4 @@ void Vector::normalize() {
 
 float Vector::length() const {
   return sqrtf(this->x * this->x + this->y * this->y + this->z * this->z);
-}
-
-Ray::Ray() = default;
-
-Ray::Ray(const Vector& origin, const Vector& direction) : origin(origin), direction(direction){};
-
-Color Ray::generateColor() const {
-  // return Color(static_cast<int>(std::abs(this->direction.x * 2)) % 255,
-  //              static_cast<int>(std::abs(this->direction.y * 2)) % 255,
-  //              static_cast<int>(std::abs(this->direction.z * 2)) % 255);
-  return Color(static_cast<int>(std::abs(this->direction.x) * 255) % 255,
-               static_cast<int>(std::abs(this->direction.y) * 255) % 255,
-               static_cast<int>(std::abs(this->direction.z) * 255) % 255);
 }
